Refine the Schedule API
