+++ OmiseSwift/Schedule.swift
@@ -1,5 +1,4 @@
 import Foundation
-import EventKit
 
 
 public struct Schedule<Data: Schedulable>: OmiseResourceObject, Equatable {
@@ -36,7 +35,7 @@ public struct Schedule<Data: Schedulable>: OmiseResourceObject, Equatable {
     public let occurrences: ListProperty<Occurrence<Data>>
     public let nextOccurrencesOnDateComponents: [DateComponents]
     
-    public let parameter: Data.Parameter
+    public let scheduleData: Data.ScheduleData
 }
 
 extension Schedule {
@@ -172,7 +171,7 @@ extension Schedule {
             }
         })
         if let parameterKey = parameterKey {
-            parameter = try container.decode(Data.Parameter.self, forKey: parameterKey)
+            scheduleData = try container.decode(Data.ScheduleData.self, forKey: parameterKey)
         } else {
             throw DecodingError.keyNotFound(
                 Schedule<Data>.CodingKeys.parameter("parameter"),
@@ -200,7 +199,7 @@ extension Schedule {
         try container.encodeOmiseDateComponents(endOnDateComponents, forKey: .endDate)
         try container.encode(occurrences, forKey: .occurrences)
         try container.encodeOmiseDateComponentsArray(nextOccurrencesOnDateComponents, forKey: .nextOccurrenceDates)
-        try container.encode(parameter, forKey: .parameter(Data.parameterKey))
+        try container.encode(scheduleData, forKey: .parameter(Data.parameterKey))
     }
 }
 
@@ -208,28 +207,6 @@ extension Calendar {
     public static let scheduleAPICalendar: Calendar = {
         return Calendar(identifier: .gregorian)
     }()
-}
-
-public protocol SchedulingParameter: Codable {}
-
-public protocol Schedulable: OmiseIdentifiableObject, OmiseCreatableObject {
-    associatedtype Parameter: SchedulingParameter
-    static func isValidParameterKey(_ key: String) -> Bool
-    static var parameterKey: String { get }
-}
-
-public protocol APISchedulable: Schedulable {
-    associatedtype ScheduleDataParams: APIJSONQuery
-}
-
-extension Schedulable {
-    public static var parameterKey: String {
-        return String(describing: self).lowercased()
-    }
-    
-    public static func isValidParameterKey(_ key: String) -> Bool {
-        return parameterKey == key
-    }
 }
 
 
@@ -275,136 +252,17 @@ extension Schedule.Status: Codable {
 }
 
 
-public struct AnySchedulable: Schedulable {
-    public enum AnySchedulingParameter: SchedulingParameter {
-        case charge(ChargeSchedulingParameter)
-        case transfer(TransferSchedulingParameter)
-        case other(json: [String: Any])
-        
-        public init(from decoder: Decoder) throws {
-            let container = try decoder.singleValueContainer()
-            
-            if let chargeParameter = try? container.decode(ChargeSchedulingParameter.self) {
-                self = .charge(chargeParameter)
-            } else if let transferParameter = try? container.decode(TransferSchedulingParameter.self) {
-                self = .transfer(transferParameter)
-            } else {
-                let json = try decoder.decodeJSONDictionary()
-                self = .other(json: json)
-            }
-        }
-        
-        public func encode(to encoder: Encoder) throws {
-            var container = encoder.singleValueContainer()
-            switch self {
-            case .charge(let parameter):
-                try container.encode(parameter)
-            case .transfer(let parameter):
-                try container.encode(parameter)
-            case .other(json: let parameter):
-                try encoder.encodeJSONDictionary(parameter)
-            }
-        }
-    }
-    
-    public typealias Parameter = AnySchedulingParameter
-    public let id: String
-    public let createdDate: Date
-    public let object: String
-    
-    public let json: [String: Any]
-    
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case createdDate = "created"
-        case object
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let json = try decoder.decodeJSONDictionary()
-        
-        guard let object = json[CodingKeys.object.stringValue] as? String else {
-            let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-            throw DecodingError.keyNotFound(CodingKeys.object, context)
-        }
-        
-        guard let id = json[CodingKeys.id.stringValue] as? String else {
-            let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-            throw DecodingError.keyNotFound(CodingKeys.id, context)
-        }
-        guard let createdDate = (json[CodingKeys.createdDate.stringValue] as? String).flatMap({
-            ISO8601DateFormatter().date(from: $0)
-        }) else {
-            let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-            throw DecodingError.keyNotFound(CodingKeys.createdDate, context)
-        }
-        
-        (self.object, self.id, self.createdDate) = (object, id, createdDate)
-        self.json = json
-    }
-    
-    public static func isValidParameterKey(_ key: String) -> Bool {
-        return key == Charge.parameterKey || key == Transfer.parameterKey
-    }
-    
-    public struct AnyScheduleData: OmiseIdentifiableObject, OmiseLiveModeObject, OmiseCreatableObject {
-        public let id: String
-        public let createdDate: Date
-        public let object: String
-        public let isLiveMode: Bool
-        
-        public let json: [String: Any]
-        
-        public init(from decoder: Decoder) throws {
-            let json = try decoder.decodeJSONDictionary()
-            
-            guard let object = json[CodingKeys.object.stringValue] as? String else {
-                let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-                throw DecodingError.keyNotFound(CodingKeys.object, context)
-            }
-            
-            guard let id = json[CodingKeys.id.stringValue] as? String else {
-                let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-                throw DecodingError.keyNotFound(CodingKeys.id, context)
-            }
-            guard let createdDate = (json[CodingKeys.createdDate.stringValue] as? String).flatMap({
-                ISO8601DateFormatter().date(from: $0)
-            }) else {
-                let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-                throw DecodingError.keyNotFound(CodingKeys.createdDate, context)
-            }
-            guard let isLiveMode = json[CodingKeys.isLiveMode.stringValue] as? Bool else {
-                let context = DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Missing object value in Schedule")
-                throw DecodingError.keyNotFound(CodingKeys.isLiveMode, context)
-            }
-            
-            (self.object, self.id, self.createdDate, self.isLiveMode) = (object, id, createdDate, isLiveMode)
-            self.json = json
-        }
-        
-        public func encode(to encoder: Encoder) throws {
-            var jsonValue = json
-            jsonValue[CodingKeys.id.stringValue] = id
-            jsonValue[CodingKeys.createdDate.stringValue] = createdDate
-            jsonValue[CodingKeys.object.stringValue] = object
-            jsonValue[CodingKeys.isLiveMode.stringValue] = isLiveMode
-            try encoder.encodeJSONDictionary(jsonValue)
-        }
-    }
-}
-
-
 extension Schedule: Listable {}
 extension Schedule: Retrievable {}
 
 
-public struct ScheduleParams<Data: APISchedulable>: APIJSONQuery {
+public struct ScheduleParams<Data: Schedulable>: APIJSONQuery {
     public let every: Int
     public let period: Period
     public let endDate: DateComponents
     public let startDate: DateComponents?
     
-    public let scheduleData: Data.ScheduleDataParams
+    public let scheduleData: Data.Parameter
     
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: Schedule<Data>.CodingKeys.self)
@@ -419,7 +277,7 @@ public struct ScheduleParams<Data: APISchedulable>: APIJSONQuery {
     }
 }
 
-extension Schedule : Creatable where Data : Creatable & APISchedulable {
+extension Schedule : Creatable where Data : Creatable & Schedulable {
     public typealias CreateParams = ScheduleParams<Data>
 }
